% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author.R
\name{author}
\alias{author}
\title{Quarto Author Schema}
\usage{
author(given, family, orcid = NULL, roles = c(), ...)
}
\arguments{
\item{given}{a character string with the author's first name(s)}

\item{family}{a character string with the author's last name(s)}

\item{orcid}{the author's unique ORCiD (see https://orcid.org/)}

\item{roles}{a vector of roles from the CRediT taxonomy (see https://casrai.org/credit/); use credit_roles() to view the full list}

\item{...}{further info to add to author object (see Details)}
}
\value{
a list
}
\description{
Set up an author object using the quarto author schema \url{https://quarto.org/docs/journals/authors.html}.
}
\details{
The author schema is as follows (the function currently does not do much checking). The metadata/roles attribute is specific to this package.

\if{html}{\out{<div class="sourceCode">}}\preformatted{author:
  - id: "LDB"
    number: 1
    name:
      given: Lisa M.
      family: DeBruine
      literal: Lisa M. DeBruine
      dropping-particle: ""
      non-dropping-particle: ""
    url: https://debruine.github.io
    email: debruine@gmail.com
    fax: ""
    orcid: 0000-0002-7523-5539
    note: ""
    acknowledgements: ""
    attributes:
      corresponding: true
      equal-contributor: false
      deceased: false
    metadata:
      roles: [Conceptualization, Methodology]
    affiliations:
    - id: UofG-SPN
      number: 1
      name: University of Glasgow
      department: School of Psychology & Neuroscience
      address: 62 Hillhead Street
      city: Glasgow
      region: Scotland
      country: United Kingdom
      postal-code: G12 8QB
      url: https://www.gla.ac.uk/schools/psychologyneuroscience/
}\if{html}{\out{</div>}}
}
\examples{
a <- author(
  "Lisa M.", "DeBruine",
  orcid = "0000-0002-7523-5539",
  roles = c("Conceptualization", "Methodology"),
  email = "debruine@gmail.com",
  affiliation = list(
    name = "University of Glasgow",
    department = "School of Psychology & Neuroscience"
  )
)
}
