% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{comma_and}
\alias{comma_and}
\title{List to Text}
\usage{
comma_and(x, comma = ", ", and = " & ", oxford = FALSE)
}
\arguments{
\item{x}{The list or vector to convert}

\item{comma}{The text to use to separate all but the last item}

\item{and}{The text to use to separate the last item}

\item{oxford}{Whether to use an oxford comma before the last item}
}
\value{
A character string
}
\description{
Convert a list or vector to text with human-readable separators, e.g., "A, B & C".
}
\examples{
comma_and(LETTERS[1:5])
comma_and(LETTERS[1:5], and = " and ")
comma_and(LETTERS[1:5], comma = "; ")

# change and to use an oxford comma
my_list <- list("Nelson Mandela",
                "an 800-year-old demigod",
                "a dildo collector")
comma_and(my_list) # probably not what you mean
comma_and(my_list, oxford = TRUE)
}
