% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\name{stripes}
\alias{stripes}
\title{Colour Stripes CSS}
\usage{
stripes(
  colors = c(pink = "#983E82", orange = "#E2A458", yellow = "#F5DC70", green = "#59935B",
    blue = "#467AAC", purple = "#61589C"),
  height = 3
)
}
\arguments{
\item{colors}{a named vector of the colours, the names are the alias and the values are the colour (in any CSS format, such as hex)}

\item{height}{the height of each stripe in pixels}
}
\value{
a character string of the CSS
}
\description{
Create the CSS for a striped border on each book page
}
\examples{
# default PsyTeachR colours
stripes()

# red, yellow and blue thick stripes
stripes(c(red = "red", yellow = "yellow", blue = "blue"), 10)
}
