% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{write_yaml}
\alias{write_yaml}
\title{Write yaml with fixed logical values}
\usage{
write_yaml(x, file)
}
\arguments{
\item{x}{the object to be converted}

\item{file}{either a character string naming a file or a connection open for writing}
}
\description{
Write yaml with fixed logical values
}
