% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{create_book}
\alias{create_book}
\title{Create a Template Book}
\usage{
create_book(
  path = "book",
  title = "Book Template",
  author = "Me",
  description = "",
  license = "CC-BY",
  google_analytics = "",
  twitter = "",
  repo_url = "",
  repo_branch = "main",
  repo_actions = "edit, issue, source",
  downloads = "pdf, epub",
  sharing = "twitter, facebook, linkedin",
  footer = paste(license, format(Sys.Date(), "\%Y"), ",", author),
  light_theme = "flatly",
  dark_theme = "darkly",
  stripe_css = stripes()
)
}
\arguments{
\item{path}{path to the location for your book on your machine}

\item{title}{title of the book}

\item{author}{author of the book}

\item{description}{short description of the book}

\item{license}{what license to add (CC-BY only for now)}

\item{google_analytics}{the google analytics address}

\item{twitter}{the twitter adddress}

\item{repo_url}{the github repo URL (e.g., "myusername/mybook")}

\item{repo_branch}{the branch to use (usually "main" or "master")}

\item{repo_actions}{github repo links to add to the right sidebar}

\item{downloads}{download types to include}

\item{sharing}{"twitter, facebook, linkedin"}

\item{footer}{defaults to "license YEAR, author"}

\item{light_theme}{"flatly"}

\item{dark_theme}{"darkly"}

\item{stripe_css}{created by stripe() function}
}
\value{
sets up a project and renders the demo
}
\description{
Create a Template Book
}
