% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{create_book}
\alias{create_book}
\title{Create a Template Book}
\usage{
create_book(
  path = "book",
  title = "Book Template",
  subtitle = "",
  authors = "Me",
  description = "Book Description",
  output_dir = "docs",
  license = "CC-BY",
  google_analytics = "",
  socials = list(),
  downloads = c("pdf", "epub"),
  sharing = c("twitter", "facebook", "linkedin"),
  margin_header = "",
  footer = paste0(license, " (", format(Sys.Date(), "\%Y"), ") ", authors),
  light_theme = "flatly",
  dark_theme = "darkly",
  css = stripes(),
  df_print = "kable",
  webexercises = TRUE,
  open = rlang::is_interactive(),
  render = TRUE
)
}
\arguments{
\item{path}{path to the location for your book on your machine}

\item{title}{title of the book}

\item{subtitle}{subtitle of the book}

\item{authors}{authors of the book (see Details)}

\item{description}{short description of the book}

\item{output_dir}{directory to render output to}

\item{license}{what license to add (CC-BY only for now)}

\item{google_analytics}{the google analytics address}

\item{socials}{a list of social media URLs to put in the footer, named as a relevant icon (see Details)}

\item{downloads}{download types to include}

\item{sharing}{"twitter, facebook, linkedin"}

\item{margin_header}{defaults to ""}

\item{footer}{defaults to "license (YEAR) author"}

\item{light_theme}{"flatly" (see https://quarto.org/docs/output-formats/html-themes.html)}

\item{dark_theme}{"darkly"}

\item{css}{custom styles (e.g., \code{stripes()} creates the signature PsyTeachR rainbow stripes)}

\item{df_print}{how to print tables (default, kable, tibble, paged)}

\item{webexercises}{whether to use webexercises for interactive exercises}

\item{open}{whether to activate the new project in RStudio}

\item{render}{whether to render the quarto book when opening}
}
\value{
sets up a project and renders the demo
}
\description{
Set up a book by creating a project at the specified path, rendering the demo book in Quarto, opening the book in a web browser, and opening the project in a new RStudio window.
}
\details{
The argument \code{authors} can be text or a structured list, e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(c("Lisa M.", "DeBruine", "0000-0002-7523-5539"),
     c("Daniel", "Lakens", "0000-0002-0247-239X"))
}\if{html}{\out{</div>}}

The argument \code{socials} adds linked icons to the right footer. See the available icons at \url{https://icons.getbootstrap.com/} E.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(mastodon = "https://tech.lgbt/@debruine",
     github = "https://github.com/debruine",
     twitter = "https://twitter.com/lisadebruine")
}\if{html}{\out{</div>}}
}
