% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_headers.R
\name{rmd2qmd}
\alias{rmd2qmd}
\title{Copy Rmd to qmd}
\usage{
rmd2qmd(from_path = "./", to_path = "./")
}
\arguments{
\item{from_path}{the source directory where your Rmd files are}

\item{to_path}{the destination directory where you want the qmd files}
}
\value{
the path to the destination directory
}
\description{
Copy Rmd files to the current and rename to .qmd
}
