% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_headers.R
\name{update_headers}
\alias{update_headers}
\title{Update headers to quarto style}
\usage{
update_headers(path = "./")
}
\arguments{
\item{path}{Path to the directory that contains your .qmd files}
}
\description{
Replaces all headers of the style "# title {#label}" with "# title {#sec-label}" and all "\@ref(label)" with "@sec-label" and all "](#label)" with "](#sec-label)".
}
